+++ src/Frontend/src/pages/mappers/subpages/MapperBackupPage.tsx
@@ -7,7 +7,7 @@ import { archiveMappers, backupMappers } from "../../../utility/fetch";
 import { MapperUpdate } from "pokeaclient";
 import { MapperFilesContext } from "../../../Contexts/availableMapperContext";
 import { OpenMapperFolderButton } from "../../../components/OpenMapperFolderButton";
-import { Advanced } from "../../../Contexts/Advanced";
+import { Advanced } from "../../../components/Advanced";
 
 export function MapperBackupPage() {
 	const filesClient = Store.client.files;